[INLONG-4939][Manager] Add version control for all the save and update APIs (#4970)
* Add version control for all the save and update APIs

* Add constraint for version field and change constant naming

+++ inlong-manager/manager-service/src/main/java/org/apache/inlong/manager/service/source/StreamSourceServiceImpl.java
@@ -252,12 +252,16 @@ public class StreamSourceServiceImpl implements StreamSourceService {
             throw new BusinessException(String.format("Source=%s is not allowed to delete", entity));
         }
 
-        entity.setVersion(entity.getVersion() + 1);
         entity.setPreviousStatus(curStatus.getCode());
         entity.setStatus(nextStatus.getCode());
         entity.setIsDeleted(id);
         entity.setModifyTime(new Date());
-        sourceMapper.updateByPrimaryKeySelective(entity);
+        int rowCount = sourceMapper.updateByPrimaryKeySelective(entity);
+        if (rowCount != InlongConstants.AFFECTED_ONE_ROW) {
+            LOGGER.error("source has already updated with groupId={}, streamId={}, name={}, curVersion={}",
+                    entity.getInlongGroupId(), entity.getInlongStreamId(), entity.getSourceName(), entity.getVersion());
+            throw new BusinessException(ErrorCodeEnum.CONFIG_EXPIRED);
+        }
         sourceFieldMapper.deleteAll(id);
 
         LOGGER.info("success to delete source for id={} by user={}", id, operator);
@@ -320,7 +324,6 @@ public class StreamSourceServiceImpl implements StreamSourceService {
         if (CollectionUtils.isNotEmpty(entityList)) {
             for (StreamSourceEntity entity : entityList) {
                 Integer id = entity.getId();
-                entity.setVersion(entity.getVersion() + 1);
                 entity.setPreviousStatus(entity.getStatus());
                 entity.setStatus(nextStatus);
                 entity.setIsDeleted(id);
